Update Program.cs
Added switches to select to process only files, only directories, or both.

NB: the /f switch has been changed to /q or /y for both consistency with most other tools and to reuse for selecting files. This means any scripts that use RR with the /f switch should be updated.